Sync up object detection code with latest year's example code
