+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/AutoDriveTest.java
@@ -5,7 +5,7 @@ package org.firstinspires.ftc.teamcode.auto;
 public abstract class AutoDriveTest extends AutoBase {
 
     public AutoDriveTest() {
-        this.doVuforia = false;
+        this.doObjectDetection = false;
     }
 }
