--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/AutoFirstBox.java
@@ -1,15 +1,0 @@
-package org.firstinspires.ftc.teamcode.auto;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-/**
- */
-//@Autonomous(name="Box First", group="AutoTest")
-public class AutoFirstBox extends AutoDriveTest {
-
-    @Override
-    public void ratCrewGo() {
-        firstBox();
-    }
-
-}
