--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/AutoSecondBox.java
@@ -1,15 +1,0 @@
-package org.firstinspires.ftc.teamcode.auto;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-/**
- */
-//@Autonomous(name="Box Second", group="AutoTest")
-public class AutoSecondBox extends AutoDriveTest {
-
-    @Override
-    public void ratCrewGo() {
-        secondBox();
-    }
-
-}
