--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/AutoThirdBox.java
@@ -1,15 +1,0 @@
-package org.firstinspires.ftc.teamcode.auto;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-/**
- */
-//@Autonomous(name="Box Third", group="AutoTest")
-public class AutoThirdBox extends AutoDriveTest {
-
-    @Override
-    public void ratCrewGo() {
-        thirdBox();
-    }
-
-}
